CtrlLib: Using beep parameter for regular prompts (fixes override problem)

+++ uppsrc/Painter/Painter.cpp
@@ -267,7 +267,7 @@ Painter& Painter::Text(double x, double y, const String& s, Font fnt, const doub
 
 Painter& Painter::Text(const Pointf& p, const char *text, Font fnt, int n, const double *dx)
 {
-	WString s = ToUtf32(text, n < 0 ? strlen(text) : n);
+	WString s = ToUtf32(text, n < 0 ? (int)strlen(text) : n);
 	return Text(p, s, fnt, s.GetCount(), dx);
 }
